requirements.txt
google-adk
google-generativeai
requests
